requirements.txt
# Core dependencies
# 使用兼容的版本
torch>=2.0.0 -i https://download.pytorch.org/whl/cpu  # 支持更新版本的 PyTorch
torchvision==0.19.0  # 与 PyTorch 匹配的 torchvision 版本
sentence-transformers==4.1.0  # 语义相似度计算
transformers>=4.26.0  # NLP 模型
fastapi>=0.68.0  # For API server
uvicorn>=0.15.0  # ASGI server for FastAPI
pydantic>=1.8.2  # Data validation for FastAPI

# Optional dependencies for advanced features
nltk>=3.6.0  # For more advanced text processing
scikit-learn>=1.0.0  # For more advanced similarity algorithms
pandas>=1.3.0  # For data manipulation
accelerate